[feature-23] Configure graphql networking
