+++ src/network/graph/client.js
@@ -1,0 +1,8 @@
+import { ApolloClient, InMemoryCache } from '@apollo/client'
+
+import { settings } from 'settings'
+
+export const make = () => new ApolloClient({
+  uri: settings().backend.graphUrl,
+  cache: new InMemoryCache(),
+})
