+++ src/network/graph/mutate.js
@@ -1,0 +1,5 @@
+import { from, of } from 'rxjs'
+import { mergeMap } from 'rxjs/operators'
+
+export const mutate = client => payload => of({})
+  .pipe(mergeMap(() => from(client.mutate(payload))))
