+++ src/network/graph/query.js
@@ -1,0 +1,5 @@
+import { from, of } from 'rxjs'
+import { mergeMap } from 'rxjs/operators'
+
+export const query = client => payload => of({})
+  .pipe(mergeMap(() => from(client.query(payload))))
